refactors main-page component
