+++ src/components/main-page/main-page.jsx
@@ -2,20 +2,17 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import {ActionCreator} from '../../store/actions/action-creator';
-import OffersList from '../offers-list/offers-list.jsx';
-import Map from '../map/map.jsx';
-import Header from '../header/header.jsx';
-import OffersSorting from '../offers-sorting/offers-sorting.jsx';
-import CitiesList from '../cities-list/cities-list.jsx';
-import MainEmpty from '../main-empty/main-empty.jsx';
-import withActiveFlag from '../../hocs/withActiveFlag/withActiveFlag';
 import NameSpace from '../../store/name-space';
 import CityPropType from '../../prop-types/city';
+import OfferPropType from '../../prop-types/offer';
+import Header from '../header/header.jsx';
+import CitiesList from '../cities-list/cities-list.jsx';
+import MainOffers from '../main-offers/main-offers.jsx';
+import MainEmpty from '../main-empty/main-empty.jsx';
 
-const OffersSortingWithActiveFlag = withActiveFlag(OffersSorting);
 
 const MainPage = (props) => {
-  const {currentOffers, cities, currentCity, sortType, onChangeCity, onSortTypeChange} = props;
+  const {currentOffers, cities, currentCity, onChangeCity} = props;
 
 
   return (
@@ -25,20 +22,9 @@ const MainPage = (props) => {
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <CitiesList cities={cities} currentCity={currentCity} onChangeCity={onChangeCity}/>
-        <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{currentOffers.length} places to stay in {currentCity.name}</b>
-              <OffersSortingWithActiveFlag activeSortType={sortType} onSortTypeChange={onSortTypeChange}/>
-              <OffersList offersCards={currentOffers} isNearByView={false}/>
-            </section>
-            <div className="cities__right-section">
-              <Map offers={currentOffers} currentCity={currentCity} isNearByView={false}/>
-            </div>
-          </div>
-        </div>
-        <MainEmpty currentCity={currentCity}/>
+        {currentOffers.length > 0
+          ? <MainOffers currentCity={currentCity} currentOffers={currentOffers}/>
+          : <MainEmpty currentCity={currentCity}/>}
       </main>
     </div>
   );
@@ -46,11 +32,9 @@ const MainPage = (props) => {
 
 MainPage.propTypes = {
   onChangeCity: PropTypes.func,
-  onSortTypeChange: PropTypes.func,
-  currentOffers: PropTypes.arrayOf(PropTypes.object).isRequired,
+  currentOffers: PropTypes.arrayOf(OfferPropType).isRequired,
   currentCity: CityPropType.isRequired,
   cities: PropTypes.arrayOf(PropTypes.string).isRequired,
-  sortType: PropTypes.string.isRequired,
 };
 
 const mapStateToProps = (state) => {
@@ -58,7 +42,6 @@ const mapStateToProps = (state) => {
     currentCity: state[NameSpace.DATA].currentCity,
     currentOffers: state[NameSpace.DATA].currentOffers,
     cities: state[NameSpace.DATA].cities,
-    sortType: state[NameSpace.DATA].sortType
   };
 };
 
@@ -67,9 +50,6 @@ const mapDispatchToProps = (dispatch) => {
     onChangeCity(city) {
       dispatch(ActionCreator.changeCity(city));
     },
-    onSortTypeChange(sortType) {
-      dispatch(ActionCreator.changeSortType(sortType));
-    }
   };
 };
